// src/app/shared/api-client/model/createOrderRequest.ts
/**
 * Pizza.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PizzaOrderRequest } from './pizzaOrderRequest';
import { SideOrderRequest } from './sideOrderRequest';


export interface CreateOrderRequest { 
    pizzas?: Array<PizzaOrderRequest> | null;
    sides?: Array<SideOrderRequest> | null;
}